# docs/contracts/openapi/backend-bff-v1.json
{
  "components": {
    "schemas": {
...
      "ProductCareResponse": {
        "additionalProperties": false,
        "properties": {
          "data_mode": {
            "enum": [
              "live",
              "replay"
            ],
            "title": "Data Mode",
            "type": "string"
          },
          "items": {
            "items": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/CareActionRecord"
                },
                {
                  "$ref": "#/components/schemas/CareFollowupRecord"
                }
              ]
            },
            "title": "Items",
            "type": "array"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          },
          "role": {
            "$ref": "#/components/schemas/ProductRole"
          },
          "schema_version": {
            "const": "product_sleep_care.v1",
            "default": "product_sleep_care.v1",
            "title": "Schema Version",
            "type": "string"
          },
          "subject_ref": {
            "minLength": 1,
            "title": "Subject Ref",
            "type": "string"
          },
          "synthetic_non_release": {
            "title": "Synthetic Non Release",
            "type": "boolean"
          }
        },
        "required": [
          "data_mode",
          "synthetic_non_release",
          "subject_ref",
          "role",
          "items"
        ],
        "title": "ProductCareResponse",
        "type": "object"
      },
...
      "ProductRole": {
        "enum": [
          "elder",
          "family",
          "doctor"
        ],
        "title": "ProductRole",
        "type": "string"
      },
...
      "sleepagent__api__product_contracts__ErrorResponse": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "minLength": 1,
            "title": "Code",
            "type": "string"
          },
          "correlation_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Correlation Id"
          },
          "data_mode": {
            "enum": [
              "live",
              "replay"
            ],
            "title": "Data Mode",
            "type": "string"
          },
          "message": {
            "minLength": 1,
            "title": "Message",
            "type": "string"
          },
          "retryable": {
            "default": false,
            "title": "Retryable",
            "type": "boolean"
          },
          "schema_version": {
            "const": "product_error.v1",
            "default": "product_error.v1",
            "title": "Schema Version",
            "type": "string"
          },
          "synthetic_non_release": {
            "title": "Synthetic Non Release",
            "type": "boolean"
          }
        },
        "required": [
          "data_mode",
          "synthetic_non_release",
          "code",
          "message"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "SleepActorAssertion": {
        "description": "Asymmetric JWS assertion bound to this HTTP request.",
        "in": "header",
        "name": "X-Sleep-Actor-Assertion",
        "type": "apiKey"
      },
      "SleepServiceCredential": {
        "description": "Rotating HTTPS Bearer credential for the calling service.",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "SleepAgent Backend",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/product/sleep/care": {
      "get": {
        "operationId": "care_product_sleep_care_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 2000,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductCareResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/sleepagent__api__product_contracts__ErrorResponse"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/sleepagent__api__product_contracts__ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/sleepagent__api__product_contracts__ErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/sleepagent__api__product_contracts__ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/sleepagent__api__product_contracts__ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/sleepagent__api__product_contracts__ErrorResponse"
                }
              }
            },
            "description": "Content Too Large"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/sleepagent__api__product_contracts__ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "501": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/sleepagent__api__product_contracts__ErrorResponse"
                }
              }
            },
            "description": "Not Implemented"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/sleepagent__api__product_contracts__ErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Care",
        "tags": [
          "Product sleep"
        ]
      }
    },
    "/product/sleep/interactions/feedback": {
      "post": {
        "operationId": "feedback_product_sleep_interactions_feedback_post",
        "parameters": [
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "title": "Idempotency-Key",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/sleepagent__api__product_contracts__FeedbackRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/sleepagent__api__product_contracts__AcceptedOperationResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/sleepagent__api__product_contracts__ErrorResponse"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/sleepagent__api__product_contracts__ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/sleepagent__api__product_contracts__ErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/sleepagent__api__product_contracts__ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/sleepagent__api__product_contracts__ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/sleepagent__api__product_contracts__ErrorResponse"
                }
              }
            },
            "description": "Content Too Large"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/sleepagent__api__product_contracts__ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "501": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/sleepagent__api__product_contracts__ErrorResponse"
                }
              }
            },
            "description": "Not Implemented"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/sleepagent__api__product_contracts__ErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Feedback",
        "tags": [
          "Product sleep"
        ]
      }
    },
    "/product/sleep/interactions/start": {
      "post": {
        "operationId": "start_product_sleep_interactions_start_post",
        "parameters": [
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "title": "Idempotency-Key",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InteractionStartRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/sleepagent__api__product_contracts__AcceptedOperationResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/sleepagent__api__product_contracts__ErrorResponse"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/sleepagent__api__product_contracts__ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/sleepagent__api__product_contracts__ErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/sleepagent__api__product_contracts__ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/sleepagent__api__product_contracts__ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/sleepagent__api__product_contracts__ErrorResponse"
                }
              }
            },
            "description": "Content Too Large"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/sleepagent__api__product_contracts__ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "501": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/sleepagent__api__product_contracts__ErrorResponse"
                }
              }
            },
            "description": "Not Implemented"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/sleepagent__api__product_contracts__ErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Start",
        "tags": [
          "Product sleep"
        ]
      }
    },
    "/product/sleep/interactions/status/{operation_id}": {
      "get": {
        "operationId": "status_product_sleep_interactions_status__operation_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "operation_id",
            "required": true,
            "schema": {
              "title": "Operation Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InteractionStatusResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/sleepagent__api__product_contracts__ErrorResponse"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/sleepagent__api__product_contracts__ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/sleepagent__api__product_contracts__ErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/sleepagent__api__product_contracts__ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/sleepagent__api__product_contracts__ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/sleepagent__api__product_contracts__ErrorResponse"
                }
              }
            },
            "description": "Content Too Large"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/sleepagent__api__product_contracts__ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "501": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/sleepagent__api__product_contracts__ErrorResponse"
                }
              }
            },
            "description": "Not Implemented"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/sleepagent__api__product_contracts__ErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Status",
        "tags": [
          "Product sleep"
        ]
      }
    },
    "/product/sleep/interactions/{interaction_id}/answer": {
      "post": {
        "operationId": "answer_product_sleep_interactions__interaction_id__answer_post",
        "parameters": [
          {
            "in": "path",
            "name": "interaction_id",
            "required": true,
            "schema": {
              "title": "Interaction Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "title": "Idempotency-Key",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InteractionAnswerRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/sleepagent__api__product_contracts__AcceptedOperationResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/sleepagent__api__product_contracts__ErrorResponse"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/sleepagent__api__product_contracts__ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/sleepagent__api__product_contracts__ErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/sleepagent__api__product_contracts__ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/sleepagent__api__product_contracts__ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/sleepagent__api__product_contracts__ErrorResponse"
                }
              }
            },
            "description": "Content Too Large"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/sleepagent__api__product_contracts__ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "501": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/sleepagent__api__product_contracts__ErrorResponse"
                }
              }
            },
            "description": "Not Implemented"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/sleepagent__api__product_contracts__ErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Answer",
        "tags": [
          "Product sleep"
        ]
      }
    },
    "/product/sleep/interactions/{interaction_id}/ask": {
      "post": {
        "operationId": "ask_product_sleep_interactions__interaction_id__ask_post",
        "parameters": [
          {
            "in": "path",
            "name": "interaction_id",
            "required": true,
            "schema": {
              "title": "Interaction Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "title": "Idempotency-Key",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InteractionAskRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/sleepagent__api__product_contracts__AcceptedOperationResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/sleepagent__api__product_contracts__ErrorResponse"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/sleepagent__api__product_contracts__ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/sleepagent__api__product_contracts__ErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/sleepagent__api__product_contracts__ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/sleepagent__api__product_contracts__ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/sleepagent__api__product_contracts__ErrorResponse"
                }
              }
            },
            "description": "Content Too Large"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/sleepagent__api__product_contracts__ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "501": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/sleepagent__api__product_contracts__ErrorResponse"
                }
              }
            },
            "description": "Not Implemented"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/sleepagent__api__product_contracts__ErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Ask",
        "tags": [
          "Product sleep"
        ]
      }
    },
...
  }
}